ati-vga: Create 2d_blt destination setup helper
A large amount of the common setup involved in a blit deals with the
destination. This moves that setup to a helper function returning a
struct (ATIBlitDest) holding all of that state. The idea here is that
this setup will be shared between blits from memory as well as from
HOST_DATA and maybe others in the future.

Otherwise this is a pure refactor of the ati_2d_blt function to make use
of the setup_2d_blt_dst helper and the struct it returns. There should be no
changes in behavior in this patch.

Signed-off-by: Chad Jablonski <chad@jablonski.xyz>