ati-vga: Implement CCE/PM4 microcode register handling
Implement registers used for loading and reading microcode for the CCE
engine. Loading the microcode is the first step for any driver
implementing CCE. Reading, while not used by drivers is very helpful for
any reverse engineering and testing work. The microcode is currently
stored but not used. This lays the groundwork for future RE work on the
microcode.

There's some quirky behavior around microcode reads that isn't
documented elsewhere. There appear to be two internal pointers, one for
reading and one for writing that can get out of sync. Comments in the
code expand on this.

Tested and validated against a Rage 128 Pro Ultra (PCI 1002:5446).

Signed-off-by: Chad Jablonski <chad@jablonski.xyz>