ati-vga: Implement PM4_STAT register and common flags
The PM4_STAT register largely mirrors the flags of the GUI_STAT
register. The exception is that the low bits of PM4_STAT reflect the
remaining slots in the CCE FIFO instead of the GUI FIFO.

This implements a very incomplete set of common flags
(MICRO_BUSY and GUI_ACTIVE) and the PM4_STAT register.

Signed-off-by: Chad Jablonski <chad@jablonski.xyz>